double instead of float for nicer API

+++ jplotter/src/main/java/hageldave/jplotter/color/ColorOperations.java
@@ -20,10 +20,10 @@ public class ColorOperations {
 	 * @param saturation
 	 * @return
 	 */
-	public static int changeSaturation(int argb, float saturation) {
+	public static int changeSaturation(int argb, double saturation) {
 		float[] hsv = new float[3];
 		Color.RGBtoHSB(Pixel.r(argb), Pixel.g(argb), Pixel.b(argb), hsv);
-		int rgb = Color.HSBtoRGB(hsv[0], hsv[1]*saturation, hsv[2]);
+		int rgb = Color.HSBtoRGB(hsv[0], (float)(hsv[1]*saturation), hsv[2]);
 		return (argb&0xff000000)|(rgb&0x00ffffff); // preserve alpha value
 	}
